[Test][Ops] move Softplus inplace regression out of AC-1 file
The activation-alignment file is the AC-1 conformance gate; its
collected-test count must match the prior PR's HEAD for parity. Move
the post-construction inplace regression to a dedicated file so the
behavioral check ships without altering AC-1 file's test count.
